refactor: 모달 타입 상수를 별도의 파일로 분리
