+++ link-namu/src/constants/modal_types.js
@@ -1,0 +1,9 @@
+const MODAL_TYPES = {
+  KakaoModal: "KakaoModal",
+  NotionModal: "NotionModal",
+  GoogleModal: "GoogleModal",
+  BookmarkAddModal: "BookmarkAddModal",
+  CategoryAddModal: "CategoryAddModal",
+};
+
+export default MODAL_TYPES;
